requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.3
Flask==3.1.3
flask-cors==6.0.2
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
packaging==26.2
Werkzeug==3.1.8
